pyproject.toml
[project]
name = "mnemosyne"
version = "0.5.0"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "zvec>=0.2.0",
    "sentence-transformers[onnx]>=5.0.0",
    "onnxruntime>=1.18.0",
    "einops>=0.8.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-ulid>=3.0.0",
    "pylate>=1.3.0",
    "numpy>=1.26.0",
    "google-genai>=1.0.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
